(WIP) Jitsu integration
Jitsu needs to know when a domU is "ready". Ideally this'd be done after
whatever userspace service is running in it has actually init'ed and
opened a socket but this will do for now. Also, write the "boot time"
in nsecs since the epoch to Xenstore to get statistics on how quickly
rumprun booted.